requirements.txt
fastapi
uvicorn[standard]
tensorflow-cpu 
Pillow
numpy
requests